- proper releases + semi-integration tests fixes
